data_clean_env/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-data-clean-env"
version = "0.1.0"
description = "Data Cleaning Environment for OpenEnv - evaluates LLMs on data quality issue identification and fixing"
requires-python = ">=3.10"
dependencies = [
    # Core OpenEnv dependencies (required for server functionality)
    "openenv-core[core]>=0.2.2",
    "fastapi>=0.115.0",
    "fastmcp>=2.0.0",
    "pydantic>=2.0.0",
    "uvicorn>=0.24.0",
    "requests>=2.31.0",
    # Data cleaning specific dependencies
    "pandas>=2.0.0",
    "numpy>=1.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
server = "data_clean_env.server.app:app"

[tool.setuptools]
packages = ["data_clean_env", "data_clean_env.server"]
package-dir = { "data_clean_env" = ".", "data_clean_env.server" = "server" }

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]
